Fix compilation warnings for 32 bit build.
git-svn-id: http://svn.cern.ch/guest/lhcb/Online/trunk@129503 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: 53b640f317949ab06a215cd17a113a6627d4c2d9

+++ Online/Online/ROMon/src/TorrentClusterLine.cpp
@@ -233,9 +233,9 @@ void TorrentClusterLine::display() {
   ::sprintf(txt,"%6d/%-6d",sum.num_pieces_done,sum.num_pieces_total);
   ::scrc_put_chars(dis,txt,sum.num_pieces_done==sum.num_pieces_total ? NORMAL : RED,line,50,0);
 
-  ::sprintf(txt,"%8ld",sum.total_download/1024/1024);
+  ::sprintf(txt,"%8d",int(sum.total_download/1024/1024));
   ::scrc_put_chars(dis,txt,NORMAL,line,64,0);
-  ::sprintf(txt,"%8ld",sum.total_upload/1024/1024);
+  ::sprintf(txt,"%8d",int(sum.total_upload/1024/1024));
   ::scrc_put_chars(dis,txt,NORMAL,line,73,0);
 
   end_update();
